git-tfs-id: [https://tfs.smartcom.software/tfs/DotNET_Project_Collection]$/TPM/ReleaseCandidate/Source;C36689

+++ Module.Frontend.TPM/Client/model/tpm/cogs/DeletedCOGS.js
@@ -18,7 +18,7 @@ Ext.define('App.model.tpm.cogs.DeletedCOGS', {
             name: 'ClientTreeObjectId', type: 'int', mapping: 'ClientTree.ObjectId',
             defaultFilterConfig: { valueField: 'ObjectId' }, breezeEntityType: 'ClientTree', hidden: false, isDefault: true
         },
-        { name: 'LVSpercent', type: 'int', hidden: false, isDefault: true }
+        { name: 'LVSpercent', type: 'float', hidden: false, isDefault: true }
     ],
     proxy: {
         type: 'breeze',
